functionality seems good for now, moving on to some styling

+++ server/db.js
@@ -128,33 +128,36 @@ const authenticate = async(credentials)=> {
 const setup = async() => {
     await client.connect();
     console.log('connected to the database');
-    const SQL = `
-        DROP TABLE IF EXISTS todos;
-        DROP TABLE IF EXISTS users;
+    if(process.env.SYNC) {
+        const SQL = `
+            DROP TABLE IF EXISTS todos;
+            DROP TABLE IF EXISTS users;
 
-        CREATE TABLE users(
-            user_id SERIAL PRIMARY KEY,
-            created_at TIMESTAMP DEFAULT now(),
-            username VARCHAR(100) UNIQUE NOT NULL,
-            password VARCHAR(100) NOT NULL
-        );
+            CREATE TABLE users(
+                user_id SERIAL PRIMARY KEY,
+                created_at TIMESTAMP DEFAULT now(),
+                username VARCHAR(100) UNIQUE NOT NULL,
+                password VARCHAR(100) NOT NULL
+            );
 
-        CREATE TABLE todos(
-            todo_id SERIAL PRIMARY KEY,
-            description VARCHAR(255),
-            user_id INTEGER REFERENCES users(user_id)
-        );
-    `;
-    await client.query(SQL);
+            CREATE TABLE todos(
+                todo_id SERIAL PRIMARY KEY,
+                description VARCHAR(255),
+                user_id INTEGER REFERENCES users(user_id)
+            );
+        `;
+        await client.query(SQL);
 
-    const [moe, lucy, ethyl] = await Promise.all([
-        createUser({ username: 'moe', password: 'm_password' }),
-        createUser({ username: 'lucy', password: 'l_password' }),
-        createUser({ username: 'ethyl', password: '1234' }),
-    ]);
-    console.log('create tables and seed data');
+        const [moe, lucy, ethyl] = await Promise.all([
+            createUser({ username: 'moe', password: 'm_password' }),
+            createUser({ username: 'lucy', password: 'l_password' }),
+            createUser({ username: 'ethyl', password: '1234' }),
+        ]);
+        
+        console.log('create tables and seed data');
+    };
+    
 };
-
 setup();
 
 module.exports = {
